requirements.txt
numpy==2.2.2
openpyxl==3.1.5
pandas==3.0.0
pyarrow==23.0.0
pyqt6==6.10.2
pyqtgraph==0.14.0
python-dotenv==1.2.1
scipy==1.15.1
yfinance==1.1.0